Search: Avoid corruption in "No results" line when using `<`
By default, when searching for things like "5 <" only the 5 is
shown and the rest of the HTML template gets sucked into the open
bracket due to it being interpolated as raw HTML.

It's be nice if the values were text-safe by default, with perhaps
a way to access the raw version explicitly if/when truly needed in
some non-HTML context. Ref https://github.com/algolia/autocomplete.js/issues/221.