Search: Make search work even if modern JS did not execute
If the browser hasn't executed the ES6 module script (e.g. not supported,
or failed to load, or still loading, or disabled by the user),
then fallback to a simple HTML form that submits to an external
search engine.

Once the ES6 module has set up the autocomplete, disable the form,
turning it effectively it back into the `<div>` we used to have.