Docs: Explain theme variables, phase out `$color-medium`
This variable is redundant. Make it easier to get started by
removing this.

Also thin-down the sample content a bit so that it's less noisy
and less an outdated copy of qunitjs.com.

Switch from greyscale defaults to a more realistic color so that
debugging is easy. I figured greyscale would make things easier,
but, meh, I don't remember why :)