Search: Add '/' (slash) keyboard shortcut for input field
Also fix placeholder styling while at it to not rely on the browser
default. The default in Firefox was quite alright, in that it seemed
to use plain white at a lower opacity which worked very well on our
purple background.

In Safari and Chrome, however, the user agent stylesheet appears to
hardcode a specific tint of grey, which completely disappeared against
our darker colored background.