Theme: Add support for blog posts
== GitHub Actions workflow ==

This commit introduces a soft dependency on the jekyll-archives plugin.
A non-blog site can continue to use Amethyst without it, in which case
any blog-specific page layout might not work correctly.

The jekyll-archive Gem package is a first-party plugin from the Jekyll
community. But, it not included in the github-pages Gem. As such, in
order to use it we have to opt-out from the default Jekyll deployment
that GitHub Pages offers, and instead deploy it as a static site that
we build ourselves from a GitHub Actions worflow (run Jekyll and
commit files to a branch). This is becoming a more common pattern in
the community already, due to github-pages also lagginb behind in
other features (and still on the Jekyll 3.x series).

== 404 ==

I have tried to most URL patterns currently used by WordPress as
configured on blog.jquery.com. But, a few URLs will likely become
404, and I noticed Amethyst did not yet have a custom 404 page.

This commit adds one, so as to ensure the visitor stays within the
site and can use the search feature (instead of ending up on a
generic Nginx or GitHub 404 error page).

== Features ==

* Blog post rendering with semantic HTML, microdata hCard, and other
  aspects that help accessible technology to understand and focus
  on the content, proper display in "Reader Mode" of Firefox and
  Safari, and SEO.

* Basic comment rendering with Gravatar.

* Index by tag.

* Index by year.

* Sidebar modules for tag list, recent posts, and yearly archives.

Not implemented:

* Pagination.

  The WordPress blog rendered full post contents on index pages,
  which made them quite long and difficult to scan.

  In Amethyst, I've rendered only excerpts on index pages, which makes
  them much more scannable I think and for the most part forgoes the
  need for pagination. There is only so much in any given year or
  category.

  For the home page, we list the 10 most recent posts, and then link
  to an even more minimal archive index where all posts are linked with
  only their title and date in a bullet list (not even an excerpt).

* Authors don't have dedicated pages.
  I considered giving authors pages like for tags, but this conflicted
  with jekyll-feed. And the data-based author index that jekyll-feed
  supports doesn't result in dedicated pages. Rather than a messy
  workaround, perhaps it's not something we need, so long as the
  author information itself is preserved.

* Archive indexes by year only, not by month.
  There typically aren't that many posts and it seemed to make for
  rather long and clunky navigation. I have compensated by making
  the archive indexes render the post excerpt only, thus making
  them very scannable.

* Comment authors are plain text (no external link for commenter URL).

* Comment dates have day-level precision (no hours/minutes/seconds).