Theme: Close nav menu when clicking outside the menu
Previously the menu would only close when clicking on the entry link
that opened the menu.

The way I fixed this is by getting rid of our custom JS approach, and
using the native `:focus-within` capability instead, since it does
this naturally already. When clicking outside the menu, the focus is
no longer on the the entry link or a sub menu item.

No checkbox hack needed, either.