Theme: Implement `docsearch:amethyst_pagetype_navigation=1`
This aids docsearch-scraper (e.g. for Typesense, Melisearch,
or Algolia) by offering a way to exclude these navigation overview
pages.

The current qunitjs.com site gets around this by luck.
It uses jekyll-algolia to index the site proactively during deployment
before the final pages are built, and the content that jekyll sees
on "group" and other overview pages is the empty string, since the
"content" there is effectively generated by the layout template in
the next phase.