Theme: Display categories atop each API page
This improves mobile browsing, where we hide the sidebar.

Previously, the sidebar was the only way to know (by eye) which category
the current page belong to, and the only way to browse the other
pages in the same category.

Now, with the "Categories" output, the categories are mentioned
explicitly, and on mobile people can click those to browe the
category tree and navigate to other nearby pages.

This is also helps with Typesense, which this way has an easy way
to find what the name of the category is for the current page.

This is inspired by the jQuery API docs which have had this kind of
category indication since 2015, and is indexed by Algolia using
the same mechanism.