Build: Fix spec.files to not include local `_config.yaml` and demo content
By default, `gem build` packages the entire directory, including the
demo's own _config.yml file. Jekyll in turn apparently discovers this
file and creates an inheritence chain between the individual site's
config data and the plugin's demo site. This in turn meant that it
was possible to disable Typesense on the example site because
within Jekyll `site.foo.bar` would read this random package's config
file and see a setting by that name, and no local override to say
otherwise.