Theme: Switch anchor line to text-decoration, enable in footer for a11y
There is a relatively new Chrome Lighthouse score penalty (Accessiblity
score is now 96 on qunitjs.com instead of 100) for
"Links rely on color to be distinguishable".

I thought this was about color contrast, but it isn't since the same
colors are fine on the left side of the footer. On the right side,
we have both text and links. Elsewhere, we have underlines for this,
so let's enable these here as well.

Switch anchor lines from border-bottom to native text-decoration
while at it. The full range of our styles can now be expresssed
natively.

https://developer.mozilla.org/en-US/docs/Web/CSS/text-decoration
https://developer.mozilla.org/en-US/docs/Web/CSS/text-underline-offset