Theme: Add `<div class=posts>` to group layout
This allows us to distinguish between normal content (at the top of a
group page like <https://qunitjs.com/api/config/>) and generated
navigation "content" (the linked heading and excerpts for each page
in the group or category).

This in turn can be used to exclude these listings from a search
index like TypeSense. We currently use

`<meta name="docsearch:amethyst_pagetype_navigation" content="1">`

for this purpose, but that's a boolean thing for the entire page.

We can't use that for <https://qunitjs.com/api/config/> which
both contains useful content and a list of pages below it, where it
is important that search suggestions link to the dedicated pages when
looking for the config properties, and not the overview page with
just a one line excerpt. However, we do want the tutorial content
above it to be discoverable through search, as otherwise there would
be no search results at all for terms like "preconfig".