Theme: Add support for author pages and username shortcuts
There was already a 'posts-author' layout for author page, but these
pages were not linked by default from anywhere. Add built-in support
for these, by linking them from the post page, and replacing the
displayed name with the author page. The author page is expected to
have the author's full name as its title, with the `post.author`
value likely being a username, or other internal ID to use as slug
in the URL.

This is merely a progressive enhancement. The default remains to
display the value as-is, and thus can continue to be used to specify
a name directly, e.g. for guest authors or for if you don't want to
create these pages.