Theme: Increase page max-width from 1040px to 1100px
At wide viewports, of the 60px extra space 50px goes to the sidebar
and 5px goes to the content area. This is motivated by improving the
sidebar to reduce wrapping of page titles, page groups, and TOC headings.

At narrow viewports, we now prioritize the content over the sidebar,
letting the sidebar shrink a bit.

Narrow 768px
* Before: 220px sidebar    , 470px content
* After:  185px sidebar (-), 505px content (+)

Wide
* Before (1040px): 220px sidebar     , 757px content
* After  (1100px): 275px sidebar (++), 762px content (+)